Обобщены алгоритмы ShadowMap
