+++ ShadowMap/inc/ShadowMap.h
@@ -1,0 +1,11 @@
+#ifndef SHADOW_MAP_H
+#define SHADOW_MAP_H
+
+#include "ShadowMapMainRenderData.h"
+
+void ShadowMap(ShadowMapMainRenderData& data);
+
+void ShadowMapDebug(ShadowMapMainRenderData& data);
+
+
+#endif // SHADOW_MAP_H
