+++ ShadowMap/inc/ShadowMapEsm.h
@@ -1,0 +1,11 @@
+#ifndef SHADOW_MAP_ESM_H
+#define SHADOW_MAP_ESM_H
+
+#include "ShadowMapEsmRenderData.h"
+
+void ShadowMapEsm(ShadowMapEsmRenderData& data);
+
+void ShadowMapEsmDebug(ShadowMapEsmRenderData& data);
+
+
+#endif // SHADOW_MAP_ESM_H
