--- ShadowMap/inc/ShadowMapEsmPerspective.h
@@ -1,11 +1,0 @@
-#ifndef SHADOW_MAP_ESM_PERSPECTIVE_H
-#define SHADOW_MAP_ESM_PERSPECTIVE_H
-
-#include "ShadowMapEsmRenderData.h"
-
-void ShadowMapEsmPerspective(ShadowMapEsmRenderData& data);
-
-void ShadowMapEsmPerspectiveDebug(ShadowMapEsmRenderData& data);
-
-
-#endif // SHADOW_MAP_ESM_PERSPECTIVE_H
