+++ ShadowMap/inc/ShadowMapPcf.h
@@ -1,0 +1,10 @@
+#ifndef SHADOW_MAP_PCF_H
+#define SHADOW_MAP_PCF_H
+
+#include "ShadowMapPcfRenderData.h"
+
+void ShadowMapPcf(ShadowMapPcfRenderData& data);
+
+void ShadowMapPcfDebug(ShadowMapPcfRenderData& data);
+
+#endif // SHADOW_MAP_PCF_H
