--- ShadowMap/inc/ShadowMapPcfPerspective.h
@@ -1,10 +1,0 @@
-#ifndef SHADOW_MAP_PCF_PERSPECTIVE_H
-#define SHADOW_MAP_PCF_PERSPECTIVE_H
-
-#include "ShadowMapPcfRenderData.h"
-
-void ShadowMapPcfPerspective(ShadowMapPcfRenderData& data);
-
-void ShadowMapPcfPerspectiveDebug(ShadowMapPcfRenderData& data);
-
-#endif // SHADOW_MAP_PCF_PERSPECTIVE_H
