--- ShadowMap/inc/ShadowMapPerspective.h
@@ -1,11 +1,0 @@
-#ifndef SHADOW_MAP_PERSPECTIVE_H
-#define SHADOW_MAP_PERSPECTIVE_H
-
-#include "ShadowMapMainRenderData.h"
-
-void ShadowMapPerspective(ShadowMapMainRenderData& data);
-
-void ShadowMapPerspectiveDebug(ShadowMapMainRenderData& data);
-
-
-#endif // SHADOW_MAP_PERSPECTIVE_H
