+++ ShadowMap/inc/ShadowMapVsm.h
@@ -1,0 +1,11 @@
+#ifndef SHADOW_MAP_VSM_H
+#define SHADOW_MAP_VSM_H
+
+#include "ShadowMapVsmRenderData.h"
+
+void ShadowMapVsm(ShadowMapVsmRenderData& data);
+
+void ShadowMapVsmDebug(ShadowMapVsmRenderData& data);
+
+
+#endif // SHADOW_MAP_VSM_H
