--- ShadowMap/inc/ShadowMapVsmPerspective.h
@@ -1,11 +1,0 @@
-#ifndef SHADOW_MAP_VSM_PERSPECTIVE_H
-#define SHADOW_MAP_VSM_PERSPECTIVE_H
-
-#include "ShadowMapVsmRenderData.h"
-
-void ShadowMapVsmPerspective(ShadowMapVsmRenderData& data);
-
-void ShadowMapVsmPerspectiveDebug(ShadowMapVsmRenderData& data);
-
-
-#endif // SHADOW_MAP_VSM_PERSPECTIVE_H
